# src/main/java/com/builderspay/config/dto/AuthDTO.java
package com.builderspay.config.dto;

public class AuthDTO {
	
	private String client_id;
	private String client_secret;

	public String getClient_id() {
		return client_id;
	}

	public void setClient_id(String client_id) {
		this.client_id = client_id;
	}

	public String getClient_secret() {
		return client_secret;
	}

	public void setClient_secret(String client_secret) {
		this.client_secret = client_secret;
	}
}